Fix: -Wconversion MdsCmprs

+++ mdsshr/MdsCmprs.c
@@ -73,17 +73,18 @@
 #include <stdlib.h>
 #include <limits.h>
 
+
 #define MIN(a,b) ((a) < (b) ? (a) : (b))
 #define MAXX 1024		/*length of longest run allowed */
 #define MAXY 32			/*maximum bits that we can pack */
 #define BITSX 10		/*number of bits for run length */
 #define BITSY 6			/*number of bits in y 0-32 */
-#define MASK(bits)  ((unsigned int)0xffffffff >> (32 - bits))
-#define YFIELD(y) ((unsigned int)(y) & MASK(BITSY))
-#define XFIELD(x) (((unsigned int)(x) & MASK(BITSX)) << BITSY)
-#define X_AND_Y(x,y) (XFIELD(x) | YFIELD(y))
-#define X_OF_INT(val) (((unsigned int)(val) >> BITSY) & MASK(BITSX))
-#define Y_OF_INT(val) ((unsigned int)(val) & MASK(BITSY))
+#define MASK(bits)   (0xffffffff >> (32 - bits))
+#define YFIELD(y)    ((unsigned int)(y) & MASK(BITSY))
+#define XFIELD(x)    (((unsigned int)(x) & MASK(BITSX)) << BITSY)
+#define X_AND_Y(x,y) (int)(XFIELD(x) | YFIELD(y))
+#define X_OF_INT(val) (int)(((unsigned int)(val) >> BITSY) & MASK(BITSX))
+#define Y_OF_INT(val) (int)((unsigned int)(val) & MASK(BITSY))
 #define SWAP_SHORTS(in,out)         ((short *)out)[0] = ((short *)in)[1],((short *)out)[1] = ((short *)in)[0]
 
 typedef struct {
@@ -213,22 +214,22 @@ Do this in runs.
     for (pn = diff, j = xn; --j >= 0; old = *p32++) {
       unsigned int delta;
       int64_t ans;
-    /**** Check for integer overflow, 
+/**** Check for integer overflow,
  *        previously was:
  *             Check for special delta of 0x80000000
  *             which is the largest negative
  *             32-bit value. On some versions of the c
  *             compiler the other <= tests do not work.
  *        if overflow, put in flag value
-      ***********/
+ ***********/
       ans = (int64_t)*p32 - (int64_t)old;
       if ((ans > (int64_t)INT_MAX) || (ans < (int64_t)-2147483647)) {
-        *pn++=ans;
+        *pn++=(int)ans;
         yy = 32;
       }
       else {
-        i = *pn++ = ans;
-	delta = (i < 0) ? -i : i;
+        i = *pn++ = (int)ans;
+	delta = (unsigned int)((i < 0) ? -i : i);
 	if (delta <= 64) {
 	  yy = signif[delta];
 	}
@@ -268,7 +269,7 @@ Do this in runs.
     **************************************************/
       tally[MAXY - 1] += tally[MAXY];
       for (ptally = tally, xsum = xn; (*ptally = xsum -= *ptally) > 0; ++ptally) ;
-      yn = ye = yy = (ptally - tally) + 1;
+      yn = ye = yy = (int)(ptally - tally) + 1;
     /*******************************************
     We have ptally pointed at last nonempty bin.
     Determine optimal breakpoint for exceptions.
@@ -288,7 +289,7 @@ Do this in runs.
   Build exception table.
   Assume 2's complement.
   *********************/
-    mark = 0xFFFFFFFF << (yn - 1);
+    mark = (int)(0xFFFFFFFF << (yn - 1));
     maxim = ~mark;
     for (pn = diff, pe = exce, j = xe; --j >= 0;) {
       while (*pn <= maxim && *pn > mark)
@@ -323,7 +324,7 @@ EXPORT int MdsXpand(int *nitems_ptr,
   int nitems = *nitems_ptr;
   char *ppack =
       memcpy(malloc(pack_dsc_ptr->arsize + 4), pack_dsc_ptr->pointer, pack_dsc_ptr->arsize);
-  int limit = pack_dsc_ptr->arsize * 8;
+  int limit = (int)pack_dsc_ptr->arsize * 8;
   int step = items_dsc_ptr->length;
   int dtype = items_dsc_ptr->dtype;
   register PF px = (PF) items_dsc_ptr->pointer;
@@ -369,7 +370,7 @@ Note the sign-extended unpacking.
       pe = exce;
       nbits = (char)ye;
       MdsUnpk(&nbits, &xe, (int *)ppack, pe, (int *)bit_ptr);
-      mark = 0xFFFFFFFF << (-yn - 1);
+      mark = (int)(0xFFFFFFFF << (-yn - 1));
     }
     else {
       pe = diff;
